refactor rest client pattern
